changes made to 09

+++ cpp09/ex00/include/BitcoinExchange.hpp
@@ -6,7 +6,7 @@
 /*   By: dmaessen <dmaessen@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/23 10:28:16 by dmaessen          #+#    #+#             */
-/*   Updated: 2024/02/28 10:07:29 by dmaessen         ###   ########.fr       */
+/*   Updated: 2024/04/15 10:10:07 by dmaessen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@
 class BitcoinExchange
 {
     private:
-        std::map<std::string, float> _m;
+        std::map<std::string, double> _m;
 
     public:
         BitcoinExchange(void);
